+++ src/innetwork-task/model/producer.h
@@ -35,7 +35,7 @@ class Producer : public Application {
         static TypeId GetTypeId (void);
         Producer();
         virtual ~Producer ();
-        void SetupProducer (uint16_t port, uint16_t itr, uint8_t rank, uint16_t vsize, std::vector<Address> &sGroup, 
+        void SetupProducer (uint16_t port, uint16_t itr, uint8_t rank, uint32_t vsize, std::vector<Address> &sGroup, 
                                         std::vector<Address> &cGroup, uint16_t basetime, std::string cc_name);
 
     protected:
